fix(mixedcontentsourceadapter.test.ts): error during delete operation of experiencepropertytype
This change is to fix the error during the delete operation for experiencepropertytype metadata
bundle. As the bundle is all JSON files and there is no XML file in the bundle there will be an
error when the delete operation fails in the backend and this is the fix for the same.