feat: Pr/875 (#885)
* fix(mixedcontentsourceadapter.ts): error - Delete operation of experiencepropertytype metadata

this change is to fix the error during delete operation for experiencepropertytype metadata api. as
the bundle is all json files and there is no xml file in place delete operation is failing when the
metadata API is failing in the backend and this is the fix for same.

* feat(mixedcontentsourceadapter.ts): error during delete operation of experiencepropertytype bundle

This change is to fix the error during the delete operation for experiencepropertytype metadata
bundle. As the bundle is all JSON files and there is no XML file in the bundle there will be an
error when the delete operation fails in the backend and this is the fix for the same.

* fix(mixedcontentsourceadapter.test.ts): error during delete operation of experiencepropertytype

This change is to fix the error during the delete operation for experiencepropertytype metadata
bundle. As the bundle is all JSON files and there is no XML file in the bundle there will be an
error when the delete operation fails in the backend and this is the fix for the same.

* test: record perf

* fix: lint fix

* test: record perf

---------

Co-authored-by: Prashanth Josyula <prashanth.16@gmail.com>
Co-authored-by: svc-cli-bot <svc_cli_bot@salesforce.com>